@@ -1,5 +1,4 @@
 ##!/usr/bin/env python
-# -*- coding: utf-8 -*-
 #
 # cassiopeia documentation build configuration file, created by
 # sphinx-quickstart on Fri Jun  9 13:47:02 2017.
@@ -20,6 +19,11 @@
 #
 import sys
 from pathlib import Path
+
+from sphinx.ext.autosummary import Autosummary
+from sphinx.ext.autosummary import get_documenter
+from docutils.parsers.rst import directives
+from sphinx.util.inspect import safe_getattr
 
 HERE = Path(__file__).parent
 sys.path[:0] = [str(HERE.parent), str(HERE / "extensions")]
@@ -50,6 +54,17 @@ extensions = [
     # "scanpydoc.autosummary_generate_imported",
     *[p.stem for p in (HERE / "extensions").glob("*.py")],
     "sphinx_gallery.load_style",
+    "myst_parser",
+    "sphinx_tabs.tabs",
+]
+
+myst_enable_extensions = [
+    "amsmath",
+    "colon_fence",
+    "deflist",
+    "dollarmath",
+    "html_image",
+    "html_admonition",
 ]
 
 # nbsphinx specific settings
@@ -66,7 +81,7 @@ templates_path = ["_templates"]
 # You can specify multiple suffix as a list of string:
 #
 # source_suffix = ['.rst', '.md']
-source_suffix = ".rst"
+source_suffix = {".rst": "restructuredtext", ".md": "markdown"}
 
 # Generate the API documentation when building
 autosummary_generate = True
@@ -85,12 +100,14 @@ annotate_defaults = True  # scanpydoc option, look into why we need this
 master_doc = "index"
 
 
-intersphinx_mapping = dict(
-    matplotlib=("https://matplotlib.org/", None),
-    numpy=("https://docs.scipy.org/doc/numpy/", None),
-    pandas=("https://pandas.pydata.org/pandas-docs/stable/", None),
-    python=("https://docs.python.org/3", None),
-)
+intersphinx_mapping = {
+    "matplotlib": ("https://matplotlib.org/", None),
+    "numpy": ("https://docs.scipy.org/doc/numpy/", None),
+    "pandas": ("https://pandas.pydata.org/pandas-docs/stable/", None),
+    "python": ("https://docs.python.org/3", None),
+    "networkx": ("https://networkx.org/documentation/stable/", None),
+    "treedata": ("https://treedata.readthedocs.io/en/latest/", None),
+}
 
 # General information about the project.
 project = "cassiopeia"
@@ -111,7 +128,7 @@ release = cassiopeia.__version__
 #
 # This is also used if you do content translation via gettext catalogs.
 # Usually you set "language" from the command line for these cases.
-language = None
+language = "en"
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -146,13 +163,13 @@ html_theme_options = {
     "twitter_url": "https://twitter.com/YosefLab",
     # "use_edit_page_button": True,
 }
-html_context = dict(
-    display_github=True,  # Integrate GitHub
-    github_user="YosefLab",  # Username
-    github_repo="Cassiopeia",  # Repo name
-    github_version="master",  # Version
-    doc_path="docs/",  # Path in the checkout to the docs root
-)
+html_context = {
+    "display_github": True,  # Integrate GitHub
+    "github_user": "YosefLab",  # Username
+    "github_repo": "Cassiopeia",  # Repo name
+    "github_version": "master",  # Version
+    "doc_path": "docs/",  # Path in the checkout to the docs root
+}
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
@@ -165,7 +182,7 @@ nbsphinx_thumbnails = {
     "notebooks/benchmark": "_static/tutorials/benchmark.png",
     "notebooks/reconstruct": "_static/tutorials/reconstruct.png",
     "notebooks/local_plotting": "_static/tutorials/local_plotting.png",
-    "notebooks/simulate_ecDNA": "_static/tutorials/local_plotting.png"
+    "notebooks/simulate_ecDNA": "_static/tutorials/local_plotting.png",
 }
 
 
@@ -185,15 +202,10 @@ mathjax_config = {
 }
 
 
-from sphinx.ext.autosummary import Autosummary
-from sphinx.ext.autosummary import get_documenter
-from docutils.parsers.rst import directives
-from sphinx.util.inspect import safe_getattr
-import re
-
 # Code for creating autosummaries for class methods / attributes
 # Taken originally from Pandas documentation
 class AutoAutoSummary(Autosummary):
+    """Autosummary directive that can document both methods and attributes."""
 
     option_spec = {
         "methods": directives.unchanged,
@@ -207,6 +219,7 @@ class AutoAutoSummary(Autosummary):
 
     @staticmethod
     def get_members(obj, typ, include_public=None):
+        """Return public members of ``obj`` matching ``typ``."""
         if not include_public:
             include_public = []
         items = []
@@ -221,6 +234,7 @@ class AutoAutoSummary(Autosummary):
         return public, items
 
     def run(self):
+        """Populate the autosummary directive with methods and attributes."""
         clazz = str(self.arguments[0])
         try:
             (module_name, class_name) = clazz.rsplit(".", 1)
@@ -230,20 +244,18 @@ class AutoAutoSummary(Autosummary):
                 _, methods = self.get_members(c, "method", ["__init__"])
 
                 self.content = [
-                    "~%s.%s" % (clazz, method)
-                    for method in methods
-                    if not method.startswith("_")
+                    f"~{clazz}.{method}" for method in methods if not method.startswith("_")
                 ]
             if "attributes" in self.options:
                 _, attribs = self.get_members(c, "attribute")
                 self.content = [
-                    "~%s.%s" % (clazz, attrib)
-                    for attrib in attribs
-                    if not attrib.startswith("_")
+                    f"~{clazz}.{attrib}" for attrib in attribs if not attrib.startswith("_")
                 ]
-        finally:
-            return super(AutoAutoSummary, self).run()
+        except (ImportError, AttributeError, ValueError):
+            return super().run()
+        return super().run()
 
 
 def setup(app):
+    """Register custom Sphinx directives used in the documentation build."""
     app.add_directive("autoautosummary", AutoAutoSummary)
